+++ projects/igniteui-angular/src/lib/combo/combo.component.spec.ts
@@ -808,7 +808,7 @@ describe('igxCombo', () => {
             combo.comboInput = {
                 value: '',
             } as any;
-            combo.filteringOptions.filterable = true;
+            combo.disableFiltering = false;
             const matchSpy = spyOn<any>(combo, 'checkMatch').and.callThrough();
             spyOn(combo.searchInputUpdate, 'emit');
 
@@ -832,7 +832,7 @@ describe('igxCombo', () => {
             expect(combo.searchInputUpdate.emit).toHaveBeenCalledTimes(2);
             expect(combo.searchInputUpdate.emit).toHaveBeenCalledWith(args);
 
-            combo.filteringOptions.filterable = false;
+            combo.disableFiltering = true;
             combo.handleInputChange();
             expect(matchSpy).toHaveBeenCalledTimes(4);
             expect(combo.searchInputUpdate.emit).toHaveBeenCalledTimes(2);
@@ -851,7 +851,7 @@ describe('igxCombo', () => {
             spyOn(mockIconService, 'addSvgIconFromText').and.returnValue(null);
             combo.ngOnInit();
             combo.data = data;
-            combo.filteringOptions.filterable = true;
+            combo.disableFiltering = false;
             combo.searchInputUpdate.subscribe((e) => {
                 e.cancel = true;
             });
@@ -1079,8 +1079,7 @@ describe('igxCombo', () => {
                 expect(combo.itemsMaxHeight).toEqual(320);
                 expect(combo.itemHeight).toEqual(32);
                 expect(combo.placeholder).toEqual('Location');
-                expect(combo.searchPlaceholder).toEqual('Enter a Search Term');
-                expect(combo.filteringOptions.filterable).toEqual(true);
+                expect(combo.disableFiltering).toEqual(false);
                 expect(combo.allowCustomValues).toEqual(false);
                 expect(combo.cssClass).toEqual(CSS_CLASS_COMBO);
                 expect(combo.type).toEqual('box');
@@ -1148,13 +1147,16 @@ describe('igxCombo', () => {
                 expect(combo.placeholder).toEqual('Location');
                 expect(combo.comboInput.nativeElement.placeholder).toEqual('Location');
 
-                expect(combo.searchPlaceholder).toEqual('Enter a Search Term');
                 expect(combo.searchInput.nativeElement.placeholder).toEqual('Enter a Search Term');
 
                 combo.searchPlaceholder = 'Filter';
                 fixture.detectChanges();
                 expect(combo.searchPlaceholder).toEqual('Filter');
                 expect(combo.searchInput.nativeElement.placeholder).toEqual('Filter');
+
+                combo.disableFiltering = true;
+                fixture.detectChanges();
+                expect(combo.searchPlaceholder).toEqual('Filter');
 
                 combo.placeholder = 'States';
                 fixture.detectChanges();
@@ -1275,9 +1277,9 @@ describe('igxCombo', () => {
                 expect(focusedItem_2.classList.contains(CSS_CLASS_FOCUSED)).toBeTruthy();
                 expect(focusedItem_1.classList.contains(CSS_CLASS_FOCUSED)).toBeFalsy();
             });
-            it(`should not render search input if both 'allowCustomValues' and 'filterable' are false`, () => {
+            it(`should not render search input if 'allowCustomValues' is false and 'disableFiltering' is true`, () => {
                 combo.allowCustomValues = false;
-                combo.filteringOptions.filterable = false;
+                combo.disableFiltering = true;
                 expect(combo.displaySearchInput).toBeFalsy();
                 combo.toggle();
                 fixture.detectChanges();
@@ -2885,7 +2887,7 @@ describe('igxCombo', () => {
                 combo.filterFunction = comboIgnoreDiacriticsFilter;
                 combo.displayKey = null;
                 combo.valueKey = null;
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: undefined };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: undefined };
                 combo.data = ['José', 'Óscar', 'Ángel', 'Germán', 'Niño', 'México', 'Méxícó', 'Mexico', 'Köln', 'München'];
                 combo.toggle();
                 fixture.detectChanges();
@@ -3160,7 +3162,7 @@ describe('igxCombo', () => {
                 combo.close(); // Filter is cleared on close
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions.filterable = false; // Filtering is disabled
+                combo.disableFiltering = true; // Filtering is disabled
                 fixture.detectChanges();
                 combo.open(); // All items are visible since filtering is disabled
                 tick();
@@ -3175,7 +3177,7 @@ describe('igxCombo', () => {
                 combo.close(); // Filter is cleared on close
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions.filterable = true; // Filtering is re-enabled
+                combo.disableFiltering = false; // Filtering is re-enabled
                 fixture.detectChanges();
                 combo.open(); // Filter is cleared on open
                 tick();
@@ -3184,13 +3186,14 @@ describe('igxCombo', () => {
             }));
             it(`should properly display "Add Item" button when filtering is off`, () => {
                 combo.allowCustomValues = true;
-                combo.filteringOptions.filterable = false;
+                combo.disableFiltering = true;
                 fixture.detectChanges();
                 expect(combo.isAddButtonVisible()).toEqual(false);
 
                 combo.toggle();
                 fixture.detectChanges();
                 expect(combo.collapsed).toEqual(false);
+                expect(combo.searchInput.nativeElement.placeholder).toEqual('Add Item');
                 const searchInput = fixture.debugElement.query(By.css(CSS_CLASS_SEARCHINPUT));
                 UIInteractions.triggerInputEvent(searchInput, combo.data[2].field);
                 fixture.detectChanges();
@@ -3240,7 +3243,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.filterFunction = (collection: any[], searchValue: any, filteringOptions: IComboFilteringOptions): any[] => {
                     if (!collection) return [];
                     if (!searchValue) return collection;
@@ -3278,7 +3281,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.filterFunction = (collection: any[], searchValue: any, filteringOptions: IComboFilteringOptions): any[] => {
                     if (!collection) return [];
                     if (!searchValue) return collection;
@@ -3315,7 +3318,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.open();
                 tick();
                 fixture.detectChanges();
@@ -3331,7 +3334,7 @@ describe('igxCombo', () => {
                 fixture.detectChanges();
                 expect(combo.dropdown.items.length).toEqual(0);
 
-                combo.filteringOptions = Object.assign({}, combo.filteringOptions, { filterable: false });
+                combo.disableFiltering = true;
                 fixture.detectChanges();
                 expect(combo.dropdown.items.length).toBeGreaterThan(0);
             }));
@@ -3659,7 +3662,7 @@ describe('igxCombo', () => {
 @Component({
     template: `
     <igx-combo #combo [placeholder]="'Location'" [data]='items'
-        [filterable]='true' [valueKey]="'field'" [groupKey]="'region'" [width]="'400px'"
+        [disableFiltering]='false' [valueKey]="'field'" [groupKey]="'region'" [width]="'400px'"
         (selectionChanging)="selectionChanging($event)" [style.--ig-size]="'var(--ig-size-' + size + ')'">
         <ng-template igxComboItem let-display let-key="valueKey">
             <div class="state-card--simple">
@@ -3742,7 +3745,7 @@ class IgxComboSampleComponent {
         </p>
         <p>
             <igx-combo #comboReactive formControlName="townCombo"
-                class="input-container" [filterable]="true" placeholder="Location(s)"
+                class="input-container" [disableFiltering]="false" placeholder="Location(s)"
                 [data]="items" [displayKey]="'field'" [groupKey]="'region'">
                 <label igxLabel>Town</label>
             </igx-combo>
@@ -3816,7 +3819,7 @@ class IgxComboFormComponent {
     <form #form="ngForm">
         <igx-combo #testCombo class="input-container" [placeholder]="'Locations'"
             name="anyName" required [(ngModel)]="values"
-            [data]="items" [filterable]="filterableFlag"
+            [data]="items" [disableFiltering]="disableFilteringFlag"
             [displayKey]="'field'" [valueKey]="'field'"
             [groupKey]="'field' ? 'region' : ''" [width]="'100%'">
             <label igxLabel>Combo Label</label>
@@ -3917,7 +3920,7 @@ export class IgxComboBindingTestComponent {
         <igx-combo #combo placeholder="Location(s)"
             [data]="citiesData"
             [allowCustomValues]="true"
-            [filterable]="true">
+            [disableFiltering]="false">
         </igx-combo>
     </div>
     `,